Updated username/password with new input IDs
The login and register forms had the same input IDs for username and password, which caused an issue particularly in the login page where both logging in and registering needed to be implemented. To remedy this conflict, I changed the input IDs of the username & password in the register form. I also updated the handlebars file for every page.